Make `blob_to_field_elements` return `Polynomial` and fix tests
